Relocate secondaryDoc write logic to useEditorAssistant
